pyproject.toml
[project]
name = "wix-scraper"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "beautifulsoup4>=4.13.4",
    "dotenv>=0.9.9",
    "google-api-python-client>=2.176.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "openai>=1.96.1",
    "pdfminer-six>=20250506",
    "pillow>=11.3.0",
    "playwright>=1.46.0",
    "playwright-stealth>=2.0.0",
    "pydantic[email]>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "tqdm>=4.67.1",
    "pre-commit>=4.2.0",
    "jinja2>=3.1.6",
    "tenacity>=9.1.2",
    "structlog>=25.4.0",
]
